requirements.txt
asgiref==3.5.2
dj-database-url==2.2.0
Django==4.0.5
django-filter==21.1
django-heroku==0.3.1
gunicorn==22.0.0
Markdown==3.6
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pygame==2.6.0
sqlparse==0.4.2
typing_extensions==4.12.2
tzdata==2022.1
whitenoise==6.7.0
